// packages/frontend/src/generated/types/AcceptanceResponse.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

import type { ArrivalType } from './ArrivalType.ts'
import type { DocumentStatus } from './DocumentStatus.ts'

/**
 * @description Response DTO for the `acceptance_document` entity.
 */
export interface AcceptanceResponse {
  /**
   * @type string
   */
  arrivalType: ArrivalType
  /**
   * @type string
   */
  createdAt: string
  /**
   * @type string, uuid
   */
  createdBy: string
  /**
   * @type string
   */
  dateAccepted: string
  /**
   * @type string,null
   */
  deletedAt?: string | null
  /**
   * @type string,null, uuid
   */
  deletedBy?: string | null
  /**
   * @type string
   */
  documentNumber: string
  /**
   * @type string,null
   */
  executedAt?: string | null
  /**
   * @type string,null, uuid
   */
  executedBy?: string | null
  /**
   * @type string, uuid
   */
  id: string
  /**
   * @type string, uuid
   */
  originDbId: string
  /**
   * @type string,null, uuid
   */
  railWaybillId?: string | null
  /**
   * @type string,null
   */
  revertedAt?: string | null
  /**
   * @type string,null, uuid
   */
  revertedBy?: string | null
  /**
   * @type string,null
   */
  sourceEntity?: string | null
  /**
   * @type string
   */
  status: DocumentStatus
  /**
   * @type string,null, uuid
   */
  transitDispatchId?: string | null
  /**
   * @type string,null, uuid
   */
  truckWaybillId?: string | null
  /**
   * @type string
   */
  updatedAt: string
  /**
   * @type string, uuid
   */
  updatedBy: string
}